Chembl compound system code Cl -> chembl.compound
also adds chembl.target

This fixes issue #16 and brings (at least these) system codes
in line with identifiers.org

It is a bit longer than "Cl" and is arguably not a "short code", but at least
now it also is a bit more recognizable.

Is this controversial/breaking change (e.g. new major version)?