Uniprot: Follow relative HTTP redirects
This fixes #52.

Basically we'll use java.net.URI as the type for our baseURL and
location instead, so that we can use its .resolve() method.

Note that now we also forbid making a IDMapperUniprot(URI)
with a base URI that is not absolute (that would cause same type
of error from HTTPClient)